python/pyproject.toml
[tool.poetry]
name = "chat bot AI"
version = "0.1.0"
description = ""
authors = ["lamxuanbao <lamxuanbao.tb@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
promptflow = "^1.10.0"
promptflow-tools = "^1.4.0"
psycopg2-binary = "^2.9.9"
tiktoken = "^0.6.0"
langchain-community = "^0.0.36"
langchain-text-splitters = "^0.0.1"
langchain-cohere = "^0.1.4"
langchain-postgres = "^0.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
create = true
in-project = true
